pyproject.toml
[project]
name = "mcp-remote"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.115.12",
    "mcp[cli]>=1.9.3",
    "python-dotenv>=1.1.0",
    "tensorflow>=2.15.0",
    "numpy>=1.24.0",
    "httpx>=0.25.0",
    "pillow>=10.0.0",
    "matplotlib>=3.10.3",
    "pandas>=2.3.1",
    "opencv-python-headless>=4.12.0.88",
    "ultralytics>=8.3.168",
]
